moved configs
