+++ src/main/java/dev/salonce/discordquizbot/infrastructure/buttons/handlers/StartNowButtonHandler.java
@@ -3,9 +3,7 @@ package dev.salonce.discordquizbot.infrastructure.buttons.handlers;
 import dev.salonce.discordquizbot.infrastructure.buttons.ButtonHandler;
 import dev.salonce.discordquizbot.infrastructure.buttons.ButtonInteractionData;
 import dev.salonce.discordquizbot.application.MatchService;
-import dev.salonce.discordquizbot.domain.MatchState;
 import discord4j.core.event.domain.interaction.ButtonInteractionEvent;
-import discord4j.core.object.entity.channel.MessageChannel;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
