+++ src/main/java/dev/salonce/discordquizbot/infrastructure/configs/ButtonHandlerChainConfig.java
@@ -1,5 +1,7 @@
-package dev.salonce.discordquizbot.infrastructure.buttons;
+package dev.salonce.discordquizbot.infrastructure.configs;
 
+import dev.salonce.discordquizbot.infrastructure.buttons.ButtonHandler;
+import dev.salonce.discordquizbot.infrastructure.buttons.ButtonHandlerChain;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
